Rename TypicalOwners to TypicalEntities in preparation for combine
