+++ src/test/java/dog/pawbook/logic/commands/DeleteOwnerCommandTest.java
@@ -2,9 +2,9 @@ package dog.pawbook.logic.commands;
 
 import static dog.pawbook.logic.commands.CommandTestUtil.assertCommandFailure;
 import static dog.pawbook.logic.commands.CommandTestUtil.assertCommandSuccess;
+import static dog.pawbook.testutil.TypicalEntities.getTypicalDatabase;
 import static dog.pawbook.testutil.TypicalIndexes.ID_FIRST_OWNER;
 import static dog.pawbook.testutil.TypicalIndexes.ID_SECOND_OWNER;
-import static dog.pawbook.testutil.TypicalOwners.getTypicalDatabase;
 import static java.util.stream.Collectors.toList;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
