+++ src/test/java/dog/pawbook/logic/commands/EditOwnerCommandTest.java
@@ -7,10 +7,10 @@ import static dog.pawbook.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
 import static dog.pawbook.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
 import static dog.pawbook.logic.commands.CommandTestUtil.assertCommandFailure;
 import static dog.pawbook.logic.commands.CommandTestUtil.assertCommandSuccess;
+import static dog.pawbook.testutil.TypicalEntities.getTypicalDatabase;
 import static dog.pawbook.testutil.TypicalIndexes.ID_FIRST_OWNER;
 import static dog.pawbook.testutil.TypicalIndexes.ID_SECOND_OWNER;
 import static dog.pawbook.testutil.TypicalIndexes.ID_THIRD_OWNER;
-import static dog.pawbook.testutil.TypicalOwners.getTypicalDatabase;
 import static java.util.stream.Collectors.toList;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
